kloROS_accuracy_stack/requirements-dev.txt
ruff==0.6.9
black==24.10.0
mypy==1.11.2
pytest==8.3.2
pip-audit==2.7.3
semgrep==1.93.0
vulture==2.11
pycln==2.4.0
PyYAML==6.0.2
